fix: add local.setup()
