+++ protocol-units/bridge/config/src/lib.rs
@@ -22,40 +22,6 @@ async fn run_all_tests() -> Result<(), anyhow::Error> {
     Ok(())
 }
 
-async fn setup() -> Result<(AnvilInstance, Child), anyhow::Error> {
-	// Initialize tracing
-	tracing_subscriber::fmt()
-		.with_env_filter(
-		EnvFilter::try_from_default_env().unwrap_or_else(|_| EnvFilter::new("info")),
-		)
-		.init();
-
-	// Get the config file
-	let dot_movement = dot_movement::DotMovement::try_from_env()?;
-	let config_file = dot_movement.try_get_or_create_config_file().await?;
-
-	// Get a matching godfig object
-	let godfig: Godfig<BridgeConfig, ConfigFile> =
-		Godfig::new(ConfigFile::new(config_file), vec!["bridge".to_string()]);
-
-	let eth_config = EthConfig::build_for_test();
-	let eth_client = EthClient::new(eth_config).await?;
-	let anvil = Anvil::new().port(eth_client.rpc_port()).spawn();
-	    
-	let (movement_client, child) = MovementClient::new_for_test(MovementConfig::build_for_test()).await?; 
-
-	// Run a godfig transaction to update the file
-	godfig
-		.try_transaction(|config| async move {
-		println!("Config: {:?}", config);
-		let (config, _) = setup().await?;
-		Ok(Some(config))
-		})
-		.await?;
-
-	Ok((anvil, child))
-}
-
 async fn testfunction1_mvt() -> Result<(), anyhow::Error> {
 	let dot_movement = dot_movement::DotMovement::try_from_env()?;
 	let config_file = dot_movement.try_get_or_create_config_file().await?;
