+++ protocol-units/bridge/setup/src/local/mod.rs
@@ -1,0 +1,65 @@
+use anyhow::{anyhow, Context};
+use commander::run_command;
+use dot_movement::DotMovement;
+use bridge_config::Config;
+
+use tracing::info;
+
+/// The local setup strategy for MCR settlement
+#[derive(Debug, Clone)]
+pub struct Local {}
+
+impl Local {
+	/// Instantiates the local setup strategy with ports on localhost
+	/// to configure for Ethernet RPC and WebSocket client access.
+	pub fn new() -> Self {
+		Self {}
+	}
+}
+
+impl Default for Local {
+	fn default() -> Self {
+		Local::new()
+	}
+}
+
+impl Local {
+        async fn setup(
+                &self, 
+                dot_movement: &DotMovement,
+		mut config: Config,
+        ) -> Result<(AnvilInstance, Child), anyhow::Error> {
+                // Initialize tracing
+                tracing_subscriber::fmt()
+                        .with_env_filter(
+                        EnvFilter::try_from_default_env().unwrap_or_else(|_| EnvFilter::new("info")),
+                        )
+                        .init();
+        
+                // Get the config file
+                let dot_movement = dot_movement::DotMovement::try_from_env()?;
+                let config_file = dot_movement.try_get_or_create_config_file().await?;
+        
+                // Get a matching godfig object
+                let godfig: Godfig<BridgeConfig, ConfigFile> =
+                        Godfig::new(ConfigFile::new(config_file), vec!["bridge".to_string()]);
+        
+                let eth_config = EthConfig::build_for_test();
+                let eth_client = EthClient::new(eth_config).await?;
+                let anvil = Anvil::new().port(eth_client.rpc_port()).spawn();
+                    
+                let (movement_client, child) = MovementClient::new_for_test(MovementConfig::build_for_test()).await?; 
+        
+                // Run a godfig transaction to update the file
+                godfig
+                        .try_transaction(|config| async move {
+                        println!("Config: {:?}", config);
+                        let (config, _) = setup().await?;
+                        Ok(Some(config))
+                        })
+                        .await?;
+        
+                Ok((anvil, child))
+        }
+
+}
